update movie and tv show resources

+++ app/src/main/java/com/example/androidjetpacksubmission/domain/Movie.kt
@@ -1,7 +1,7 @@
 package com.example.androidjetpacksubmission.domain
 
 data class Movie(
-    val movieId: Short,
+    val movieId: Int,
     val movieTitle: String,
     val moviePoster: Int,
     val movieOverview: String,
